+++ xpilot/src/client/xinit.h
@@ -80,7 +80,7 @@ extern void Talk_resize(void);
 extern void Talk_cursor(bool visible);
 extern void Talk_map_window(bool map);
 extern int Talk_do_event(XEvent *event);
-extern int Talk_paste(char* data, int len, bool overwrite);
+extern int Talk_paste(char* data, size_t len, bool overwrite);
 extern int Talk_place_cursor(XButtonEvent *xbutton, bool pending);
 extern void Talk_window_cut(XButtonEvent *xbutton);
 extern void Talk_cut_from_messages(XButtonEvent *xbutton);
